Fields table
There's something strange going on with how the fields table is being populated in Kicad.  I found several components without the value field filled in, but if I clicked in the box, and said OK, then the field would get updated from the schematic.  In any case, they should be filled in now.  I added a folder to hold copies of the datasheets until we have a central place/method to store them.  I also added some information for other fields.  These fields should be populated in the library rather than at the schematic level.