Removed CAL connection on AX5043
The CAL connection is used to calibrate an offset, which we're not planning on needing.  Plus, I am seeing an odd coupling path on the dev board which is causing noise on the main VCO.  The datasheet recommends grounding the GPADC lines if they're not used.
So, I removed the connection and grounded the ADCs..