Multiple changes
Added series resistors and adjusted zones
Moved J1 to be closer to the edge.  This is needed because if it's not the right distance, then the cable shroud interferes.
Added extra bypass (C17 was supposed to be for clock) where C17 used to be.
Modified silk as needed
Modified the ISSP connector supply to be 3V3_protected
Modified port net on 3V3 latch up circuit to be 3V3_P to try to avoid confusing net names