Update SAMD21.kicad_sch
Various notes inserted.  I have a concern that we might need hardware flow control on the data serial port.  We might be able to use software flow control, or some other out of band method if it turns out that TCP overflows the input buffer, and doesn't cleanly back off.  We also might consider looking for ways to tweak TCP like limiting the max Frame size and Window.  The current buffer size on the data port is only 8k, but thats still smaller than a typical TCP frame. (or so i think ;)